requirements.txt
pymodbus==3.7.0
paho-mqtt
pyserial
python-can
pytz>=2023.3
python-disk-collections
